src/test/org/intranet/elevator/model/operate/controller/CarAssignmentsTest.java
/*
 * Copyright 2003 Neil McKellar and Chris Dailey
 * All rights reserved.
 */
package org.intranet.elevator.model.operate.controller;

import java.util.ArrayList;
import java.util.List;

import junit.framework.TestCase;

import org.intranet.elevator.model.Floor;
import org.intranet.sim.event.EventQueue;

/**
 * @author Neil McKellar and Chris Dailey
 *
 */
public class CarAssignmentsTest extends TestCase
{
  private EventQueue eQ;
  private List<Floor> floors;
  private Floor floor1;
  private Floor floor2;
  private Floor floor3;
  private Floor floor4;
  private Assignment base;
  private Assignment up1;
  private Assignment up2;
  private Assignment up3;
  private Assignment up4;
  private Assignment down1;
  private Assignment down2;
  private Assignment down3;
  private Assignment down4;

  public CarAssignmentsTest(String name)
  {
    super(name);
  }

  protected void setUp() throws Exception
  {
    super.setUp();
    
    eQ = new EventQueue();

    floor1 = new Floor(eQ, 1, 0, 10);
    floor2 = new Floor(eQ, 2, 10, 20);
    floor3 = new Floor(eQ, 3, 20, 30);
    floor4 = new Floor(eQ, 4, 30, 40);
    
    floors = new ArrayList<Floor>();
    floors.add(floor1);
    floors.add(floor2);
    floors.add(floor3);
    floors.add(floor4);
    
    base = new Assignment(floor1, Direction.NONE);
    
    up1 = new Assignment(floor1, Direction.UP);
    up2 = new Assignment(floor2, Direction.UP);
    up3 = new Assignment(floor3, Direction.UP);
    up4 = new Assignment(floor4, Direction.UP);
    
    down1 = new Assignment(floor1, Direction.DOWN);
    down2 = new Assignment(floor2, Direction.DOWN);
    down3 = new Assignment(floor3, Direction.DOWN);
    down4 = new Assignment(floor4, Direction.DOWN);
  }
  
  final public void testCarAssignments()
  {
    CarAssignments a = new CarAssignments();
    assertNotNull(a);
  }
  
  final public void testAddToNull()
  {
    CarAssignments a = new CarAssignments();
    a.addAssignment(floors, base, up1);
    
    // confirm that list contains up1
    assertTrue(a.contains(up1));
    // confirm that it *only* contains up1
    List<Assignment> assignments = a.getAssignments();
    assertEquals(1, assignments.size());
    assertEquals(up1, assignments.get(0));
  }
  
  final public void testAddTwice()
  {
    CarAssignments a = new CarAssignments();
    a.addAssignment(floors, base, up1);
    // confirm that list contains up1
    assertTrue(a.contains(up1));
    // add it again - should have no effect
    a.addAssignment(floors, base, up1);
    // confirm that it *only* contains up1 (once)
    List<Assignment> assignments = a.getAssignments();
    assertEquals(1, assignments.size());
    assertEquals(up1, assignments.get(0));
  }
  
  final public void testAddNoneUpUp()
  {
    // base = 1NONE
    // add: 3UP, 2UP
    // list order should be: 2UP, 3UP
    CarAssignments a = new CarAssignments();
    a.addAssignment(floors, base, up3);
    // add in front of up3
    a.addAssignment(floors, base, up2);
    List<Assignment> assignments = a.getAssignments();
    assertEquals(2, assignments.size());
    assertEquals(up2, assignments.get(0));
    assertEquals(up3, assignments.get(1));
  }
  
  final public void testAddNoneDownDown()
  {
    // betcha this one doesn't work
    // maxStops() only checks base.Direction and math is biased to UP
    
    // base = 1NONE
    // add: 2DOWN, 3DOWN
    // list order should be: 3DOWN, 2DOWN
    CarAssignments a = new CarAssignments();
    a.addAssignment(floors, base, down2);
    // add in front of down2
    a.addAssignment(floors, base, down3);
    List<Assignment> assignments = a.getAssignments();
    assertEquals(2, assignments.size());
    assertEquals(down3, assignments.get(0));
    assertEquals(down2, assignments.get(1));
  }
  
  final public void testAddDownUpDown()
  {
    // base =  3DOWN
    // add: 4UP, 2DOWN
    // list order should be: 2DOWN, 4UP
    CarAssignments a = new CarAssignments();
    a.addAssignment(floors, down3, up4);
    // add in front of up4
    a.addAssignment(floors, down3, down2);
    List<Assignment> assignments = a.getAssignments();
    assertEquals(2, assignments.size());
    assertEquals(down2, assignments.get(0));
    assertEquals(up4, assignments.get(1));
  }
  
  final public void testAddUpDownUp()
  {
    // base: 2UP
    // add: 1DOWN, 3UP
    // list order should be: 3UP, 1DOWN
    CarAssignments a = new CarAssignments();
    a.addAssignment(floors, up2, down1);
    // add in front of down1
    a.addAssignment(floors, up2, up3);
    List<Assignment> assignments = a.getAssignments();
    assertEquals(2, assignments.size());
    assertEquals(up3, assignments.get(0));
    assertEquals(down1, assignments.get(1));
  }
  
  final public void testAddDownUpUp()
  {
    // base = 3DOWN
    // add: 2UP, 1UP
    // list order should be: 1UP, 2UP
    CarAssignments a = new CarAssignments();
    a.addAssignment(floors, down3, up2);
    // add in front of up2
    a.addAssignment(floors, down3, up1);
    List<Assignment> assignments = a.getAssignments();
    assertEquals(2, assignments.size());
    assertEquals(up1, assignments.get(0));
    assertEquals(up2, assignments.get(1));
  }
  
  final public void testAddDownDownDown()
  {
    // base = 3DOWN
    // add: 1DOWN, 2DOWN
    // list order should be: 2DOWN, 1DOWN
    CarAssignments a = new CarAssignments();
    a.addAssignment(floors, down3, down1);
    // add in front of down1
    a.addAssignment(floors, down3, down2);
    List<Assignment> assignments = a.getAssignments();
    assertEquals(2, assignments.size());
    assertEquals(down2, assignments.get(0));
    assertEquals(down1, assignments.get(1));
  }
  
  final public void testAddUpUpUp()
  {
    // base = 1UP
    // add: 3UP, 2UP
    // list order should be: 2UP, 3UP
    CarAssignments a = new CarAssignments();
    a.addAssignment(floors, up1, up3);
    // add in front of up3
    a.addAssignment(floors, up1, up2);
    List<Assignment> assignments = a.getAssignments();
    assertEquals(2, assignments.size());
    assertEquals(up2, assignments.get(0));
    assertEquals(up3, assignments.get(1));
  }
  
  final public void testAddUpDownDown()
  {
    // base = 1UP
    // add: 2DOWN, 3DOWN
    // list order should be: 3DOWN, 2DOWN
    CarAssignments a = new CarAssignments();
    a.addAssignment(floors, up1, down2);
    // add in front of down2
    a.addAssignment(floors, up1, down3);
    List<Assignment> assignments = a.getAssignments();
    assertEquals(2, assignments.size());
    assertEquals(down3, assignments.get(0));
    assertEquals(down2, assignments.get(1));
  }
  
  final public void testTopEqualsBase()
  {
    // math in maxStops() checks value of topFloor when base.Direction = DOWN
    // this test confirms that the math is correct
    CarAssignments a = new CarAssignments();
    a.addAssignment(floors, down4, down2);
    // add in front of down2
    a.addAssignment(floors, down4, down3);
    List<Assignment> assignments = a.getAssignments();
    assertEquals(2, assignments.size());
    assertEquals(down3, assignments.get(0));
    assertEquals(down2, assignments.get(1));
  }
  
  final public void testTopEqualsDest()
  {
    // math in maxStops() checks value of topFloor when dest.Direction = DOWN
    // this test confirms that math is correct
    CarAssignments a = new CarAssignments();
    a.addAssignment(floors, up2, down3);
    // add in front of down3
    a.addAssignment(floors, up2, down4);
    List<Assignment> assignments = a.getAssignments();
    assertEquals(2, assignments.size());
    assertEquals(down4, assignments.get(0));
    assertEquals(down3, assignments.get(1));
  }
  
  final public void testWrapAroundUp()
  {
    // base = 3UP
    // add: 4UP, 2UP, 3DOWN, 1UP
    // list order should be: 4UP, 3DOWN, 1UP, 2UP
    CarAssignments a = new CarAssignments();
    a.addAssignment(floors, up3, up4);
    a.addAssignment(floors, up3, up2);
    // add in front of up2
    a.addAssignment(floors, up3, down3);
    a.addAssignment(floors, up3, up1);
    List<Assignment> assignments = a.getAssignments();
    assertEquals(4, assignments.size());
    assertEquals(up4, assignments.get(0));
    assertEquals(down3, assignments.get(1));
    assertEquals(up1, assignments.get(2));
    assertEquals(up2, assignments.get(3));
  }
  
  final public void testWrapAroundDown()
  {
    // base = 2DOWN
    // add: 1DOWN, 3DOWN, 2UP, 4DOWN
    // list order should be: 1DOWN, 2UP, 4DOWN, 3DOWN
    CarAssignments a = new CarAssignments();
    a.addAssignment(floors, down2, down1);
    a.addAssignment(floors, down2, down3);
    // add in front of down3
    a.addAssignment(floors, down2, up2);
    a.addAssignment(floors, down2, down4);
    List<Assignment> assignments = a.getAssignments();
    assertEquals(4, assignments.size());
    assertEquals(down1, assignments.get(0));
    assertEquals(up2, assignments.get(1));
    assertEquals(down4, assignments.get(2));
    assertEquals(down3, assignments.get(3));
  }
  
  final public void testGetCurrent()
  {
    CarAssignments a = new CarAssignments();
    assertEquals(null, a.getCurrentAssignment());
    a.addAssignment(floors, base, up1);
    assertEquals(up1, a.getCurrentAssignment());
  }
  
  final public void testIncluding()
  {
    // test that iteratorIncluding() includes new assignment
    CarAssignments assign = new CarAssignments();
    assign.addAssignment(floors, base, up1);
    CarAssignments a = assign.getCarAssignmentsIncluding(floors, base, up2);
    List<Assignment> assignments = a.getAssignments();
    assertEquals(2, assignments.size());
    assertEquals(up1, assignments.get(0));
    assertEquals(up2, assignments.get(1));
  }
  
  final public void testRemove()
  {
    // add: 1UP, 2UP
    // after remove: *only* 2UP
    CarAssignments a = new CarAssignments();
    a.addAssignment(floors, base, up1);
    a.addAssignment(floors, base, up2);
    assertTrue(a.contains(up1));
    assertTrue(a.contains(up2));
    a.removeAssignment(up1);
    assertFalse(a.contains(up1));
    assertTrue(a.contains(up2));
  }
}
